Update part1b.md
change the Chinese translation in line 568.
The origin text say : "One situation leading to the disappearance of _this_ arises". It means that in following situation, 'this' disappears.
The current Chinese version "一种消除这种由'this'所引起的问题的方法就是", means "one way to eliminate problems caused by the disappearance of 'this' is...", which is the opposite of the original meaning.